refactor: extract swap stage state machine to swap_state.rs
Move the SwapState<T, Stage> generic wrapper, all Stage structs (Stage1-5,
StageFailed), the GroupSwapState trait and its implementations, and all
SwapState impl blocks (including Stage4::check_for_price_violation) out
of lib.rs into a dedicated swap_state.rs module.

Internal types SwapGroupPair and GroupSwapState are pub(crate). Methods
that were previously private to pub mod pallet are pub(crate) to remain
accessible from the execution helpers.
